Move to jest test runner instead of custom mocha runner
