+++ api_tests/.gitignore
@@ -3,3 +3,4 @@ log
 *.js
 gen/
 yarn-error.log
+dist/
