+++ api_tests/.prettierignore
@@ -1,1 +1,2 @@
 gen/
+dist/
