--- api_tests/dry/config.toml
@@ -1,1 +1,0 @@
-actors = ["alice", "bob", "charlie"]
