--- api_tests/e2e/rfc003/btc_eth-erc20/config.toml
@@ -1,2 +1,0 @@
-ledgers = ["bitcoin", "ethereum"]
-actors = ["alice", "bob"]
