+++ api_tests/lib/actor_test.ts
@@ -1,16 +1,17 @@
 import { Actors } from "./actors";
 import { createActors } from "./create_actors";
+import JasmineSmacker from "smack-my-jasmine-up";
 import { timeout } from "./utils";
 
 /*
  * This test function will take care of instantiating the actors and tearing them down again
  * after the test, regardless if the test succeeded or failed.
  */
-function nActorTest(
-    name: string,
+async function nActorTest(
     actorNames: ["alice", "bob", "charlie"] | ["alice", "bob"] | ["alice"],
     testFn: (actors: Actors) => Promise<void>
 ) {
+    const name = JasmineSmacker.getCurrentTestName();
     if (!name.match(/[A-z0-9\-]+/)) {
         // We use the test name as a file name for the log and hence need to restrict it.
         throw new Error(
@@ -18,52 +19,42 @@ function nActorTest(
         );
     }
 
-    it(name, async function() {
-        this.timeout(100_000); // absurd timeout. we have our own one further down
-        const actors = await createActors(name, actorNames);
+    const actors = await createActors(name, actorNames);
 
-        try {
-            await timeout(60_000, testFn(actors));
-        } catch (e) {
-            for (const actorName of actorNames) {
-                await actors.getActorByName(actorName).dumpState();
-            }
-            throw e;
-        } finally {
-            for (const actorName of actorNames) {
-                actors.getActorByName(actorName).stop();
-            }
+    try {
+        await timeout(60_000, testFn(actors));
+    } catch (e) {
+        for (const actorName of actorNames) {
+            await actors.getActorByName(actorName).dumpState();
         }
-    });
+        throw e;
+    } finally {
+        for (const actorName of actorNames) {
+            actors.getActorByName(actorName).stop();
+        }
+    }
 }
 
 /*
  * Instantiates a new e2e test based on three actors
  *
  */
-export function threeActorTest(
-    name: string,
+export async function threeActorTest(
     testFn: (actors: Actors) => Promise<void>
 ) {
-    nActorTest(name, ["alice", "bob", "charlie"], testFn);
+    await nActorTest(["alice", "bob", "charlie"], testFn);
 }
 
 /*
  * Instantiates a new e2e test based on two actors
  */
-export function twoActorTest(
-    name: string,
-    testFn: (actors: Actors) => Promise<void>
-) {
-    nActorTest(name, ["alice", "bob"], testFn);
+export async function twoActorTest(testFn: (actors: Actors) => Promise<void>) {
+    await nActorTest(["alice", "bob"], testFn);
 }
 
 /*
  * Instantiates a new e2e test based on one actor
  */
-export function oneActorTest(
-    name: string,
-    testFn: (actors: Actors) => Promise<void>
-) {
-    nActorTest(name, ["alice"], testFn);
+export async function oneActorTest(testFn: (actors: Actors) => Promise<void>) {
+    await nActorTest(["alice"], testFn);
 }
