+++ api_tests/lib/cnd/cnd_instance.ts
@@ -66,15 +66,6 @@ export class CndInstance {
             );
         }
 
-        this.process.on("exit", (code: number, signal: number) => {
-            if (global.verbose) {
-                console.log(
-                    `cnd ${this.actorConfig.name} exited with ${code ||
-                        `signal ${signal}`}`
-                );
-            }
-        });
-
         const logReader = new LogReader(logFile);
         await logReader.waitForLogMessage("Starting HTTP server on");
 
@@ -83,6 +74,9 @@ export class CndInstance {
     }
 
     public stop() {
+        if (global.verbose) {
+            console.log(`terminating cnd ${this.actorConfig.name}`);
+        }
         this.process.kill("SIGINT");
         this.process = null;
     }
