+++ api_tests/lib/ledgers/ethereum.ts
@@ -1,3 +1,4 @@
 export interface EthereumNodeConfig {
     rpc_url: string;
+    tokenContract: string;
 }
