+++ api_tests/package.json
@@ -6,8 +6,8 @@
     "scripts": {
         "check": "tsc && prettier --check '**/*.{ts,json,yml}' && tslint --project .",
         "postinstall": "mkdir -p ./gen && json2ts -i ./siren.schema.json -o ./gen/siren.d.ts",
-        "pretest": "cargo build --bin cnd",
-        "test": "ts-node ./harness.ts",
+        "pretest": "cargo build --bin cnd && tsc",
+        "test": "jest",
         "fix": "tslint --project . --fix && prettier --write '**/*.{ts,js,json,yml}'"
     },
     "engines": {
@@ -15,6 +15,24 @@
     },
     "author": "CoBloX Team",
     "license": "ISC",
+    "jest": {
+        "preset": "ts-jest",
+        "roots": [
+            "./tests"
+        ],
+        "testRegex": "\\.ts$",
+        "transform": {
+            "^.+\\.(t|j)s$": "ts-jest"
+        },
+        "moduleFileExtensions": [
+            "ts",
+            "js",
+            "json",
+            "node"
+        ],
+        "testEnvironment": "./dist/src/test_environment",
+        "testTimeout": 180000
+    },
     "dependencies": {
         "@iarna/toml": "^2.2.3",
         "@radar/lnrpc": "^0.9.0-beta",
@@ -25,14 +43,16 @@
         "@types/chai-subset": "^1.3.3",
         "@types/dockerode": "^2.5.24",
         "@types/glob": "^7.1.1",
+        "@types/jest": "^25.1.3",
+        "@types/jsdom": "^16.1.0",
         "@types/log4js": "^2.3.5",
-        "@types/mocha": "^7.0.2",
         "@types/node": "^13.9",
         "@types/rimraf": "^2.0.3",
         "@types/scrypt": "^6.0.0",
         "@types/tail": "^2.0.0",
         "@types/tempfile": "^3.0.0",
         "@types/urijs": "^1.19.6",
+        "async-mutex": "^0.1.4",
         "bcoin": "https://github.com/bcoin-org/bcoin#2496acc7a98a43f00a7b5728eb256877c1bbf001",
         "bignumber.js": "^9.0.0",
         "bitcoin-core": "^3.0.0",
@@ -50,21 +70,23 @@
         "ethers": "^4.0.45",
         "get-port": "^5.1.1",
         "glob": "^7.1.6",
+        "jest": "^25.1.0",
         "js-sha256": "^0.9.0",
         "json-schema-to-typescript": "^8.1.0",
         "ln-service": "^47.15.2",
         "log4js": "^6.1.2",
-        "mocha": "^7.1.0",
         "multiaddr": "^7.4.0",
         "pem-ts": "^2.0.0",
         "prettier": "^1.19.1",
         "readline-promise": "^1.0.4",
         "rimraf": "^3.0.2",
         "satoshi-bitcoin": "^1.0.4",
+        "smack-my-jasmine-up": "^0.0.3",
         "tail": "^2.0.3",
         "temp-write": "^4.0.0",
         "testcontainers": "^2.6.0",
         "tmp": "^0.1.0",
+        "ts-jest": "^25.2.1",
         "ts-node": "^8.6.2",
         "tslint": "^6.0.0",
         "tslint-config-prettier": "^1.18.0",
