--- api_tests/sanity/basic_lnd_payment.ts
@@ -1,17 +1,0 @@
-import { twoActorTest } from "../lib/actor_test";
-import { AssetKind } from "../lib/asset";
-import { LedgerKind } from "../lib/ledgers/ledger";
-
-setTimeout(function() {
-    twoActorTest("sanity-lnd-alice-pays-bob", async function({ alice, bob }) {
-        await alice.sendRequest(
-            { ledger: LedgerKind.Lightning, asset: AssetKind.Bitcoin },
-            { ledger: LedgerKind.Bitcoin, asset: AssetKind.Bitcoin }
-        );
-        const { rHash, paymentRequest } = await bob.lnCreateInvoice("20000");
-        await alice.lnPayInvoiceWithRequest(paymentRequest);
-        await bob.lnAssertInvoiceSettled(rHash);
-    });
-
-    run();
-}, 0);
