--- api_tests/sanity/config.toml
@@ -1,2 +1,0 @@
-ledgers = ["bitcoin"]
-actors = ["alice", "bob"]
