+++ api_tests/types/smack-my-jasmine-up/index.d.ts
@@ -1,0 +1,5 @@
+declare module "smack-my-jasmine-up" {
+    export default class JasmineSmacker {
+        static getCurrentTestName();
+    }
+}
